+++ src/tools/sessionStore.ts
@@ -1,3 +1,6 @@
+import { AndroidUiautomator2Driver } from 'appium-uiautomator2-driver';
+import { XCUITestDriver } from 'appium-xcuitest-driver';
+
 let driver: any = null;
 let sessionId: string | null = null;
 
@@ -13,3 +16,9 @@ export function getDriver() {
 export function getSessionId() {
   return sessionId;
 }
+
+export const getPlatformName = (driver: any): string => {
+  if (driver instanceof AndroidUiautomator2Driver) return 'Android';
+  if (driver instanceof XCUITestDriver) return 'iOS';
+  throw new Error('Unknown driver type');
+};
